修改部分文件

+++ calendar/XGYCalendar/XGYCalendarScrollView.m
@@ -248,7 +248,7 @@ static NSString *const kCellIdentifier = @"cell";
         
         // 当前月
         if (indexPath.row >= firstWeekday && indexPath.row < firstWeekday + totalDays) {
-            cell.todayLabel.text = [NSString stringWithFormat:@"%ld", indexPath.row - firstWeekday + 1];
+            cell.todayLabel.text = [NSString stringWithFormat:@"%d", (int)(indexPath.row - firstWeekday + 1)];
             cell.todayLabel.textColor = kEveryDayAndEveryYearTextColor;
             
             // 标识今天
@@ -272,11 +272,11 @@ static NSString *const kCellIdentifier = @"cell";
         // 补上前后月的日期，淡色显示
         else if (indexPath.row < firstWeekday) {
             int totalDaysOflastMonth = [self.monthArray[3] intValue];
-            cell.todayLabel.text = [NSString stringWithFormat:@"%ld", totalDaysOflastMonth - (firstWeekday - indexPath.row) + 1];
+            cell.todayLabel.text = [NSString stringWithFormat:@"%d", (int)(totalDaysOflastMonth - (firstWeekday - indexPath.row) + 1)];
             cell.todayLabel.textColor = [UIColor colorWithWhite:0.85 alpha:1.0];
             cell.todayCircle.backgroundColor = [UIColor clearColor];
         } else if (indexPath.row >= firstWeekday + totalDays) {
-            cell.todayLabel.text = [NSString stringWithFormat:@"%ld", indexPath.row - firstWeekday - totalDays + 1];
+            cell.todayLabel.text = [NSString stringWithFormat:@"%d", (int)(indexPath.row - firstWeekday - totalDays + 1)];
             cell.todayLabel.textColor = [UIColor colorWithWhite:0.85 alpha:1.0];
             cell.todayCircle.backgroundColor = [UIColor clearColor];
         }
@@ -292,7 +292,7 @@ static NSString *const kCellIdentifier = @"cell";
         
         // 当前月
         if (indexPath.row >= firstWeekday && indexPath.row < firstWeekday + totalDays) {
-            cell.todayLabel.text = [NSString stringWithFormat:@"%ld", indexPath.row - firstWeekday + 1];
+            cell.todayLabel.text = [NSString stringWithFormat:@"%d", (int)(indexPath.row - firstWeekday + 1)];
             cell.todayLabel.textColor = kEveryDayAndEveryYearTextColor;
             cell.userInteractionEnabled = YES;
             
@@ -318,12 +318,12 @@ static NSString *const kCellIdentifier = @"cell";
         else if (indexPath.row < firstWeekday) {
             XGYCalendarMonth *lastMonthInfo = self.monthArray[0];
             NSInteger totalDaysOflastMonth = lastMonthInfo.totalDays;
-            cell.todayLabel.text = [NSString stringWithFormat:@"%ld", totalDaysOflastMonth - (firstWeekday - indexPath.row) + 1];
+            cell.todayLabel.text = [NSString stringWithFormat:@"%d", (int)(totalDaysOflastMonth - (firstWeekday - indexPath.row) + 1)];
             cell.todayLabel.textColor = [UIColor colorWithWhite:0.85 alpha:1.0];
             cell.todayCircle.backgroundColor = [UIColor clearColor];
             cell.userInteractionEnabled = NO;
         } else if (indexPath.row >= firstWeekday + totalDays) {
-            cell.todayLabel.text = [NSString stringWithFormat:@"%ld", indexPath.row - firstWeekday - totalDays + 1];
+            cell.todayLabel.text = [NSString stringWithFormat:@"%d", (int)(indexPath.row - firstWeekday - totalDays + 1)];
             cell.todayLabel.textColor = [UIColor colorWithWhite:0.85 alpha:1.0];
             cell.todayCircle.backgroundColor = [UIColor clearColor];
             cell.userInteractionEnabled = NO;
@@ -339,7 +339,7 @@ static NSString *const kCellIdentifier = @"cell";
         // 当前月
         if (indexPath.row >= firstWeekday && indexPath.row < firstWeekday + totalDays) {
             
-            cell.todayLabel.text = [NSString stringWithFormat:@"%ld", indexPath.row - firstWeekday + 1];
+            cell.todayLabel.text = [NSString stringWithFormat:@"%d", (int)(indexPath.row - firstWeekday + 1)];
             cell.todayLabel.textColor = kEveryDayAndEveryYearTextColor;
             
             // 标识今天
@@ -362,11 +362,11 @@ static NSString *const kCellIdentifier = @"cell";
         else if (indexPath.row < firstWeekday) {
             XGYCalendarMonth *lastMonthInfo = self.monthArray[1];
             NSInteger totalDaysOflastMonth = lastMonthInfo.totalDays;
-            cell.todayLabel.text = [NSString stringWithFormat:@"%ld", totalDaysOflastMonth - (firstWeekday - indexPath.row) + 1];
+            cell.todayLabel.text = [NSString stringWithFormat:@"%d", (int)(totalDaysOflastMonth - (firstWeekday - indexPath.row) + 1)];
             cell.todayLabel.textColor = [UIColor colorWithWhite:0.85 alpha:1.0];
             cell.todayCircle.backgroundColor = [UIColor clearColor];
         } else if (indexPath.row >= firstWeekday + totalDays) {
-            cell.todayLabel.text = [NSString stringWithFormat:@"%ld", indexPath.row - firstWeekday - totalDays + 1];
+            cell.todayLabel.text = [NSString stringWithFormat:@"%d", (int)(indexPath.row - firstWeekday - totalDays + 1)];
             cell.todayLabel.textColor = [UIColor colorWithWhite:0.85 alpha:1.0];
             cell.todayCircle.backgroundColor = [UIColor clearColor];
         }
